# Evm Stack: dup'n'swap optimized