This PR optimizes `DUP` and `SWAP` instructions and replaces the work in #2818 .

When working on #2818 I reviewed JITted ASM and found that `PushZero` and similar methods are pretty nicely inlined by JIT itself. This results in a much smaller PR, that uses `Word` structure just internally, as all the other cases provided similar results. As it's visible for Dup'n'Swap case, gains are pretty nice.

I did run `Ethereum.Blockchain.Block.Test` suite, to ensure that I don't introduce havoc like in the previous take on this.

### Setup

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.1379 (1909/November2018Update/19H2)
Intel Core i7-6700HQ CPU 2.60GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.103
  [Host]        : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  .NET Core 5.0 : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Job=.NET Core 5.0  Runtime=.NET Core 5.0  

```

### Before

|   Method |                    v |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|--------- |--------------------- |----------:|----------:|----------:|------:|------:|------:|----------:|----------:|
|     Swap |                    ? | 21.254 ns | 0.4208 ns | 0.5010 ns |     - |     - |     - |         - |    1226 B |
|      Dup |                    ? | 12.100 ns | 0.2358 ns | 0.2090 ns |     - |     - |     - |         - |    1229 B |

### After

|   Method |                    v |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|--------- |--------------------- |----------:|----------:|----------:|------:|------:|------:|----------:|----------:|
|     Swap |                    ? | 10.489 ns | 0.1409 ns | 0.1782 ns |     - |     - |     - |         - |     688 B |
|      Dup |                    ? | 10.055 ns | 0.1822 ns | 0.1704 ns |     - |     - |     - |         - |     764 B |