@tkstanczak 2021-03-07
Hi @Scooletz we just need to find a way of running all tests on an external PR and then we can merge it :)