@tkstanczak 2021-03-07
@matilote any idea on how to execute test suite on external PRs?