@Scooletz 2021-03-07
@matilote I'm not sure which suite should be run and how considitions are defined, but we could introduce a label that would be used to conditionally run a job(s) for an external PR like in the https://stackoverflow.com/questions/59588605/how-to-check-for-a-label-in-a-github-action-condition